feat: Add edit view for expenses
This commit adds a new view called `edit` to the `myapp` application. The `edit` view allows users to edit an expense by providing a form to update the expense details. When the form is submitted, the expense is updated in the database. This feature enhances the functionality of the application by enabling users to modify their expenses.