feat: Add django.contrib.humanize to INSTALLED_APPS
This commit adds the 'django.contrib.humanize' app to the 'INSTALLED_APPS' list in the 'settings.py' file. By including this app, we enable the use of humanize filters and tags in our Django templates. This feature enhances the user experience by providing convenient formatting options for numbers, dates, and file sizes.